test: support for spanish, french and portugues greeting
