+++ hello/hello_test.go
@@ -1,0 +1,48 @@
+package main
+
+import "testing"
+
+func TestHello(t *testing.T) {
+    assertCorrectMessage := func(t testing.TB, got, want string)  {
+        t.Helper()
+
+		if got != want {
+			t.Errorf("got %q want %q", got, want)
+		}
+    }
+
+	t.Run("saying hello to people", func(t *testing.T) {		
+		got := Hello("Matheus", "")
+		want := "Hello, Matheus"
+	
+        assertCorrectMessage(t, got, want)
+	})
+
+    t.Run("say 'Hello, World' when an empty string is supplied", func(t *testing.T) {
+        got:= Hello("", "")
+        want := "Hello, World"
+
+        assertCorrectMessage(t, got, want)
+    })
+
+    t.Run("in Spanish", func(t *testing.T) {
+        got := Hello("Elodie", "Spanish")
+        want := "Hola, Elodie"
+
+        assertCorrectMessage(t, got, want)
+    })
+
+    t.Run("in French", func(t *testing.T) {
+        got := Hello("Karim", "French")
+        want := "Bonjour, Karim"
+
+        assertCorrectMessage(t, got, want)
+    })
+
+    t.Run("in Portuguese", func(t *testing.T) {
+        got := Hello("Amanda", "Portuguese")
+        want := "Olá, Amanda"
+
+        assertCorrectMessage(t, got, want)
+    })
+}
